+++ docs/scripts/views/snackbar.vue
@@ -5,28 +5,92 @@
       <p>Transient popup notifications.</p>
     </div>
 
-    // TODO
+    <h4>Toast</h4>
     <div class="snippet-demo">
-      <ui-button effect @click.native="show">show</ui-button>
-      <ui-snackbar :message="message" :active="active" @done="reset"></ui-snackbar>
+      <ui-button raised effect @click.native="show('toast')">Show</ui-button>
+      <ui-snackbar :active="toast.active" :message="toast.message" @done="hide('toast')"></ui-snackbar>
     </div>
+    <ui-code language="html" :code="code1.html"></ui-code>
+    <ui-code language="js" :code="code1.js"></ui-code>
+
+    <h4>snackbar</h4>
+    <div class="snippet-demo">
+      // TODO
+    </div>
+
+    <h4>Snackbar API</h4>
+    <ui-tabs effect position="left" :active="tab" @switched="onChange">
+      <ui-panel tab="props">
+        <ui-table
+          class="docs-table"
+          :data="docs.props.data"
+          :thead="docs.props.thead"
+          :tbody="docs.props.tbody">
+        </ui-table>
+      </ui-panel>
+      <ui-panel tab="slots">
+        <ui-table
+          class="docs-table"
+          :data="docs.slots.data"
+          :thead="docs.slots.thead"
+          :tbody="docs.slots.tbody">
+        </ui-table>
+      </ui-panel>
+      <ui-panel tab="events">
+        <ui-table
+          class="docs-table"
+          :data="docs.events.data"
+          :thead="docs.events.thead"
+          :tbody="docs.events.tbody">
+        </ui-table>
+      </ui-panel>
+    </ui-tabs>
   </div>
 </template>
 
 <script>
+import snackbarDocs from '../apidocs/snackbar';
+
 export default {
   data() {
     return {
-      message: 'Hello BalmUI',
-      active: false
+      toast: {
+        active: false,
+        message: 'Hello BalmUI'
+      },
+      code1: {
+        html: require('../snippets/snackbar/demo1-template.html'),
+        js: require('../snippets/snackbar/demo1-script.html')
+      },
+      tab: 0,
+      docs: {
+        props: {
+          data: snackbarDocs.props,
+          thead: this.$docs.props.thead,
+          tbody: this.$docs.props.tbody
+        },
+        slots: {
+          data: snackbarDocs.slots,
+          thead: this.$docs.slots.thead,
+          tbody: this.$docs.slots.tbody
+        },
+        events: {
+          data: snackbarDocs.events,
+          thead: this.$docs.events.thead,
+          tbody: this.$docs.events.tbody
+        }
+      }
     };
   },
   methods: {
-    show() {
-      this.active = true;
+    show(type) {
+      this[type].active = true;
     },
-    reset() {
-      this.active = false;
+    hide(type) {
+      this[type].active = false;
+    },
+    onChange(tab) {
+      this.tab = tab;
     }
   }
 };