+++ docs/scripts/views/menu.vue
@@ -8,7 +8,7 @@
     // TODO
     <ui-menu effect name="menu1" :data="menus"></ui-menu>
 
-    <ui-menu effect name="menu2">
+    <ui-menu name="menu2">
       <template scope="props">
         <li v-for="menu in menus" :class="props.className">{{ menu.name }}</li>
       </template>