+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.10 beta
+
+### restore dist
+
 ## 0.1.9 beta
 
 ### Update docs